+++ CHANGELOG.md
@@ -10,6 +10,9 @@
         - Fix `Custom CSS / JS` Atom having bad HTML with non-existing file path (#1105)
         - Forcing Mobile Menu Items to always display full width no matter the breakpoint (thanks @JoomFX - #1109)
         - Fixed zIndex issue in Mobile Menu in Firefox and IE (thanks @JoomFX - #1109)
+2. [Joomla](#joomla)
+    1. [](#new)
+        - Template installer: Copy configuration for new outlines
 3. [WordPress](#wordpress)
     3. [](#bugfix)
         -  Internal Error in admin Settings tab when there are no menus (#1102)